sstv_decoder.py
import os
import time
import numpy as np
from scipy.io import wavfile
import logging
from PIL import Image
import io

# Setup logging
logger = logging.getLogger(__name__)

# This function would normally use the PySSTVpp or pySSTV library
# Since we're implementing this directly, we'll create a simplified SSTV decoder
def decode_sstv(wav_file_path, output_folder, unique_id):
    """
    Decode SSTV signal from a WAV file
    
    Args:
        wav_file_path: Path to the WAV file
        output_folder: Folder to save the decoded image
        unique_id: Unique identifier for the output file
        
    Returns:
        tuple: (sstv_mode, image_path, processing_time)
    """
    start_time = time.time()
    
    try:
        # Read the WAV file
        logger.info(f"Reading WAV file: {wav_file_path}")
        sample_rate, audio_data = wavfile.read(wav_file_path)
        
        # Ensure audio data is in the right format
        if len(audio_data.shape) > 1:
            # Take the first channel if stereo
            audio_data = audio_data[:, 0]
        
        # Analyze the audio to detect SSTV mode
        sstv_mode = detect_sstv_mode(audio_data, sample_rate)
        
        if not sstv_mode:
            logger.error("No SSTV mode detected in the audio file")
            return None, None, 0
            
        logger.info(f"Detected SSTV mode: {sstv_mode}")
        
        # Decode the SSTV signal
        image_data = process_sstv_signal(audio_data, sample_rate, sstv_mode)
        
        if image_data is None:
            logger.error("Failed to decode SSTV signal")
            return None, None, 0
            
        # Save the decoded image
        output_filename = f"{unique_id}_decoded.png"
        output_path = os.path.join(output_folder, output_filename)
        
        # Convert image_data to PIL Image and save
        image = Image.fromarray(image_data.astype('uint8'), 'RGB')
        image.save(output_path)
        
        processing_time = time.time() - start_time
        logger.info(f"SSTV decoding completed in {processing_time:.2f} seconds")
        
        return sstv_mode, output_path, processing_time
        
    except Exception as e:
        logger.error(f"Error decoding SSTV signal: {str(e)}")
        raise

def detect_sstv_mode(audio_data, sample_rate):
    """
    Detect the SSTV mode from the audio data
    
    In a real implementation, this would analyze the VIS (Vertical Interval Signaling) code
    to determine the SSTV mode. For simplicity, we'll return a default mode.
    """
    # Simplified mode detection
    # In a real implementation, you would analyze the header/sync signals
    
    # For demonstration purposes, we'll do a very basic check
    # Check for the presence of sync pulses at expected intervals
    
    # Let's assume Martin M1 mode for now (popular SSTV mode)
    # In a real implementation, we would analyze the VIS code
    
    # Basic check for signal strength and presence of sync pulses
    if len(audio_data) > sample_rate * 2:  # At least 2 seconds of audio
        # Check if there's a strong enough signal
        signal_strength = np.max(np.abs(audio_data)) > 0.1 * np.max(np.abs(audio_data))
        
        if signal_strength:
            # For simplicity, detect if it's Scottie, Martin, or Robot
            # This is a very simplified detection that wouldn't work in reality
            # but serves as a placeholder for the actual detection logic
            
            # Check frequency of pulses in first few seconds
            sync_intervals = detect_sync_intervals(audio_data[:int(sample_rate * 5)], sample_rate)
            
            if 0.06 < sync_intervals < 0.08:
                return "Scottie S1"
            elif 0.08 < sync_intervals < 0.12:
                return "Martin M1"
            elif 0.03 < sync_intervals < 0.05:
                return "Robot 36"
            else:
                return "Unknown SSTV Mode"
    
    return None

def detect_sync_intervals(audio_segment, sample_rate):
    """
    Detect the average interval between sync pulses
    This is a simplified version that wouldn't work in practice
    but represents the concept
    """
    # In real implementation, we would do proper signal processing
    # Including filtering, demodulation and sync detection
    
    # For demo purposes, just return a value in the Martin M1 range
    return 0.1  # Martin M1 has sync pulses roughly every 0.1 seconds

def process_sstv_signal(audio_data, sample_rate, sstv_mode):
    """
    Process SSTV signal and convert to image
    
    In a real implementation, this would do the actual signal demodulation
    and conversion to an image according to the specific SSTV mode.
    """
    # This is a simplified example that wouldn't work in practice
    # In a real-world application, we would use proper SSTV decoding libraries
    
    # For demonstration purposes, create a simple test image
    # In real implementation, this would be replaced with actual decoding logic
    
    # Set image dimensions based on SSTV mode
    if "Scottie" in sstv_mode:
        width, height = 320, 256
    elif "Martin" in sstv_mode:
        width, height = 320, 240
    elif "Robot" in sstv_mode:
        width, height = 320, 240
    else:
        width, height = 320, 240
    
    # Create a blank RGB image
    image = np.zeros((height, width, 3), dtype=np.uint8)
    
    try:
        # Simplified "decoding" for demonstration
        # In reality, we'd use PySSTVpp or pySSTV for proper decoding
        
        # This would normally extract pixel data from the demodulated signal
        # For demo purposes, we'll create a gradient pattern based on audio characteristics
        
        # Extract some basic characteristics from the audio
        audio_abs = np.abs(audio_data)
        audio_normalized = audio_abs / np.max(audio_abs) if np.max(audio_abs) > 0 else audio_abs
        
        # Use audio characteristics to create a pattern
        for y in range(height):
            for x in range(width):
                # Create a gradient pattern
                r = int((x / width) * 255)
                g = int((y / height) * 255)
                
                # Use some audio features for the blue channel
                audio_idx = int((y * width + x) / (width * height) * len(audio_normalized))
                if audio_idx < len(audio_normalized):
                    b = int(audio_normalized[audio_idx] * 255)
                else:
                    b = 0
                
                image[y, x] = [r, g, b]
        
        # Add some "sync" lines to simulate SSTV pattern
        for y in range(0, height, 20):
            image[y, :, :] = [255, 255, 255]
        
        return image
    
    except Exception as e:
        logger.error(f"Error in SSTV signal processing: {str(e)}")
        return None
